use the same stringify function
