+++ src/components/Cell.tsx
@@ -2,6 +2,7 @@ import { asyncBufferFromUrl, parquetMetadataAsync } from 'hyparquet'
 import { useEffect, useState } from 'react'
 import type { FileSource } from '../lib/sources/types.js'
 import { parquetDataFrame } from '../lib/tableProvider.js'
+import { stringify } from '../lib/utils.js'
 import Breadcrumb, { BreadcrumbConfig } from './Breadcrumb.js'
 import Layout from './Layout.js'
 
@@ -74,31 +75,4 @@ export default function CellView({ source, row, col, config }: CellProps) {
       <pre className="viewer text">{text}</pre>
     </Layout>
   )
-}
-
-/**
- * Robust stringification of any value, including json and bigints.
- */
-function stringify(value: unknown): string {
-  if (typeof value === 'string') return value
-  if (typeof value === 'number') return value.toLocaleString('en-US')
-  if (Array.isArray(value)) {
-    return `[\n${value.map((v) => indent(stringify(v), 2)).join(',\n')}\n]`
-  }
-  if (value === null || value === undefined) return JSON.stringify(value)
-  if (value instanceof Date) return value.toISOString()
-  if (typeof value === 'object') {
-    return `{${Object.entries(value)
-      .filter((d) => d[1] !== undefined)
-      .map(([k, v]) => `${k}: ${stringify(v)}`)
-      .join(', ')}}`
-  }
-  return '{}'
-}
-
-function indent(text: string | undefined, spaces: number) {
-  return text
-    ?.split('\n')
-    .map((line) => ' '.repeat(spaces) + line)
-    .join('\n')
 }
