+++ src/components/viewers/CellPanel.tsx
@@ -1,5 +1,6 @@
-import { DataFrame, stringify } from 'hightable'
+import { DataFrame } from 'hightable'
 import { useEffect, useState } from 'react'
+import { stringify } from '../../lib/utils.js'
 import ContentHeader from './ContentHeader.js'
 
 interface ViewerProps {
@@ -36,8 +37,7 @@ export default function CellPanel({ df, row, col, setProgress, setError, onClose
         if (asyncCell === undefined) {
           throw new Error(`Cell missing at column ${columnName}`)
         }
-        /* TODO(SL): use the same implementation of stringify, here and in Cell.tsx */
-        const text = await asyncCell.then(cell => stringify(cell as unknown) ?? '{}')
+        const text = await asyncCell.then(stringify)
         setText(text)
       } catch (error) {
         setError(error as Error)
